Adding wasm support setup
