+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["sim-core", "sim-app"]
+members = ["sim-core", "sim-app", "sim-web"]
 resolver = "2"
 
 [workspace.package]
@@ -15,3 +15,8 @@ num-complex = "0.4"
 noise = "0.9"
 rand = "0.8"
 palette = "0.7"
+
+# Optimize WASM builds for size
+[profile.release]
+opt-level = 's'
+lto = true
