+++ sim-app/Cargo.toml
@@ -6,17 +6,29 @@ authors.workspace = true
 license.workspace = true
 description = "An interactive Mandelbrot-style visualizer built with egui"
 
+[lib]
+name = "sim_app"
+path = "src/lib.rs"
+
 [[bin]]
 name = "complex-systems-visualizer"
 path = "src/main.rs"
 
+[features]
+default = ["native"]
+native = ["eframe/default", "sim-core/parallel"]
+web = ["eframe/glow"]
+
 [dependencies]
-sim-core = { path = "../sim-core" }
-eframe = "0.29"
+sim-core = { path = "../sim-core", default-features = false }
+eframe = { version = "0.29", default-features = false }
 egui = "0.29"
 egui_plot = "0.29"
-rayon.workspace = true
 num-complex.workspace = true
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+# Only include rayon for native builds
+rayon.workspace = true
 
 # --- macOS bundle metadata for cargo-bundle ---
 [package.metadata.bundle]
