+++ sim-web/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "sim-web"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+description = "WebAssembly version of the Complex Systems Visualizer"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+sim-app = { path = "../sim-app", default-features = false, features = ["web"] }
+eframe = { version = "0.29", default-features = false, features = ["glow"] }
+wasm-bindgen = "0.2"
+wasm-bindgen-futures = "0.4"
+console_error_panic_hook = "0.1"
+log = "0.4"
+
+# Web APIs for DOM access
+web-sys = { version = "0.3", features = ["Window", "Document", "Element", "HtmlCanvasElement"] }
+
+# Enable JS feature for getrandom (required for WASM)
+getrandom = { version = "0.2", features = ["js"] }
